Azure_Function/z-fa-dss-eventingestion/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-servicebus==7.0.0b8
azure-storage-blob
azure-identity
pytz